04-flask-app/requirements.txt
Flask==2.1.0
python-dotenv == 1.0.0
pytest==7.4.0
gunicorn==20.1.0
